Split package structure
